Move example objects out of DimensionReduction
