+++ src/main/scala/DimensionReduction/DimensionStatUtil.scala
@@ -25,8 +25,3 @@ object DimensionStatUtil {
 
 }
 
-object RandomDistTest extends App {
-  println(DimensionStatUtil.getAverageMinDistanceInHypercube(4, 147, 1))
-  println(DimensionStatUtil.getAverageMinDistanceInHypercube(4, 147, 10))
-  println(DimensionStatUtil.getAverageMinDistanceInHypercube(4, 147, 100))
-}
