+++ src/main/scala/DimensionReduction/LinearUtil.scala
@@ -274,59 +274,3 @@ object LinearUtil {
     }
   }
 }
-
-object LinearUtilTest extends App {
-  val a: Point = Point(Vector(0.9051310475138654, 0.08442008172760171, -0.416666577286393))
-  val b: Point = Point(Vector(0.9278894781448365, 0.03992321446481126, -0.3707118197399426))
-  val c: Point = Point(Vector(0.9192682247682198, 0.13235814155761652, -0.3707118197399426))
-
-  println((b-a).get.toString)
-  println((c-a).get.toString)
-
-
-  println(LinearUtil.getRREF(Vector((b-a).get.map(Rational(_)), (c-a).get.map(Rational(_)))))
-
-}
-
-object InverseTest extends App {
-
-  val m: Vector[Vector[Rational]] = Vector(
-    Vector(Rational(1.0), Rational(2.0), Rational(3.0)),
-    Vector(Rational(6.0), Rational(5.0), Rational(4.0)),
-    Vector(Rational(1.0), Rational(0.0), Rational(0.0))
-  )
-
-  LinearUtil.getInverse(m) match {
-    case Some(inverse) => inverse foreach println
-    case None => println("Uh oh! Matrix wasn't computed...")
-  }
-
-}
-
-object MatrixMultTest extends App {
-  val m: Vector[Vector[Rational]] = Vector(
-    Vector(Rational(1.0), Rational(2.0), Rational(3.0)),
-    Vector(Rational(6.0), Rational(5.0), Rational(4.0)),
-    Vector(Rational(1.0), Rational(0.0), Rational(0.0))
-  )
-
-  LinearUtil.getInverse(m) match {
-    case Some(inverse) => LinearUtil.matrixMult(m, inverse) foreach println
-    case None => println("Uh oh! Matrix wasn't computed...")
-  }
-}
-
-
-object ProjectionTest extends App {
-
-  val basis: Vector[Vector[Rational]] = Vector(Vector(Rational(2.0), Rational(1.0)))
-
-  val P = LinearUtil.getProjectionMatrix(basis)
-
-  P foreach println
-
-  LinearUtil.matrixMult(P.get, basis.transpose) foreach println
-
-  LinearUtil.matrixMult(P.get, Vector(Vector(Rational(1.0), Rational(-2.0))).transpose) foreach println
-
-}
